+++ Tunny/Enum/GcAfterTrial.cs
@@ -1,0 +1,9 @@
+namespace Tunny.Enum
+{
+    public enum GcAfterTrial
+    {
+        Always,
+        HasGeometry,
+        NoExecute,
+    }
+}