+++ Tunny/Enum/GrasshopperStates.cs
@@ -1,0 +1,9 @@
+namespace Tunny.Enum
+{
+    public enum GrasshopperStates
+    {
+        RequestSent,
+        RequestProcessing,
+        RequestProcessed
+    }
+}