+++ Tunny/Enum/StorageType.cs
@@ -1,0 +1,11 @@
+namespace Tunny.Enum
+{
+    public enum StorageType
+    {
+        InMemory,
+        Sqlite,
+        Postgres,
+        MySql,
+        Journal,
+    }
+}