+++ Tunny/Handler/OptimizeLoop.cs
@@ -3,15 +3,13 @@ using System.Collections.Generic;
 using System.ComponentModel;
 using System.Linq;
 
-using Grasshopper.Kernel;
-
 using Tunny.Component.Optimizer;
+using Tunny.Enum;
+using Tunny.Input;
 using Tunny.PostProcess;
-using Tunny.PreProcess;
 using Tunny.Settings;
 using Tunny.Solver;
 using Tunny.Type;
-using Tunny.UI;
 
 namespace Tunny.Handler
 {
@@ -41,9 +39,9 @@ namespace Tunny.Handler
 
             if (s_component != null)
             {
-                s_component.OptimizationWindow.GrasshopperStatus = OptimizationWindow.GrasshopperStates.RequestSent;
+                s_component.OptimizationWindow.GrasshopperStatus = GrasshopperStates.RequestSent;
                 s_worker?.ReportProgress(100, pState);
-                while (s_component.OptimizationWindow.GrasshopperStatus != OptimizationWindow.GrasshopperStates.RequestProcessed)
+                while (s_component.OptimizationWindow.GrasshopperStatus != GrasshopperStates.RequestProcessed)
                 { /* just wait until the cows come home */ }
             }
 
@@ -53,7 +51,7 @@ namespace Tunny.Handler
         private static double[] RunOptimizationLoop(BackgroundWorker worker)
         {
             List<Variable> variables = s_component.GhInOut.Variables;
-            List<IGH_Param> objectives = s_component.GhInOut.Objectives;
+            Objective objectives = s_component.GhInOut.Objectives;
             Dictionary<string, FishEgg> enqueueItems = s_component.GhInOut.EnqueueItems;
             bool hasConstraint = s_component.GhInOut.HasConstraint;
 
@@ -70,17 +68,15 @@ namespace Tunny.Handler
 
         private static TrialGrasshopperItems EvaluateFunction(ProgressState pState, int progress)
         {
-            s_component.OptimizationWindow.GrasshopperStatus = OptimizationWindow.GrasshopperStates.RequestSent;
+            s_component.OptimizationWindow.GrasshopperStatus = GrasshopperStates.RequestSent;
 
             s_worker.ReportProgress(progress, pState);
-            while (s_component.OptimizationWindow.GrasshopperStatus != OptimizationWindow.GrasshopperStates.RequestProcessed)
+            while (s_component.OptimizationWindow.GrasshopperStatus != GrasshopperStates.RequestProcessed)
             { /*just wait*/ }
 
-            TunnyObjective objective = s_component.GhInOut.GetObjectiveValues();
             return new TrialGrasshopperItems
             {
-                ObjectiveValues = objective.Numbers,
-                ObjectiveImages = objective.Images,
+                Objectives = s_component.GhInOut.Objectives,
                 GeometryJson = s_component.GhInOut.GetGeometryJson(),
                 Attribute = s_component.GhInOut.GetAttributes(),
                 Artifacts = s_component.GhInOut.Artifacts,