+++ Tunny/PostProcess/TrialGrasshopperItems.cs
@@ -1,16 +1,14 @@
 using System.Collections.Generic;
-using System.Drawing;
 
-using Tunny.PreProcess;
+using Tunny.Input;
 
 namespace Tunny.PostProcess
 {
     public class TrialGrasshopperItems
     {
-        public double[] ObjectiveValues { get; set; }
+        public Objective Objectives { get; set; }
         public string[] GeometryJson { get; set; }
         public Dictionary<string, List<string>> Attribute { get; set; }
-        public Bitmap[] ObjectiveImages { get; set; }
         public Artifact Artifacts { get; set; }
     }
 }