+++ Tunny/Solver/OptimizationHandlingInfo.cs
@@ -13,7 +13,8 @@ namespace Tunny.Solver
         public dynamic ArtifactBackend { get; set; }
         public Dictionary<string, FishEgg> EnqueueItems { get; set; }
         public string[] ObjectiveNames { get; set; }
-        public HumanInTheLoop HumanInTheLoop { get; set; }
+        public HumanInTheLoop.HumanSliderInput HumanSliderInput { get; set; }
+        public HumanInTheLoop.Preferential Preferential { get; set; }
 
         public OptimizationHandlingInfo(int nTrials, double timeout, dynamic study, dynamic storage,
                                         dynamic artifactBackend, Dictionary<string, FishEgg> enqueueItems,