+++ Tunny/Solver/Visualize.cs
@@ -3,6 +3,7 @@ using System.Windows.Forms;
 
 using Python.Runtime;
 
+using Tunny.Enum;
 using Tunny.PostProcess;
 using Tunny.Settings;
 using Tunny.UI;