+++ Tunny/UI/OptimizeWindowTab/FileTab.cs
@@ -3,7 +3,7 @@ using System.Diagnostics;
 using System.IO;
 using System.Windows.Forms;
 
-using Tunny.Settings;
+using Tunny.Enum;
 using Tunny.Storage;
 using Tunny.Util;